reorganized the defineCommand test
now it just holds specific `defineCommand` test code and test code related with lua context or lua utils specifics was put in the `test/lua.js` file.
